requirements.txt
streamlit==1.28.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.18.0
